+++ src/Riichi/Efficiency.hs
@@ -1,0 +1,17 @@
+{- |
+Module      : Riichi.Efficiency
+Description : Functions for calculating discard tile efficiency
+License     : BSD-3-Clause
+Maintainer  : surplussinewaves@gmail.com
+-}
+module Riichi.Efficiency where
+
+import Riichi.Meld
+import Riichi.Shanten
+import Riichi.Tile
+
+optimalDiscard :: Hand -> Tile
+optimalDiscard = undefined
+
+improvements :: Hand -> [Tile]
+improvements = undefined